app/requirements.txt
Flask
mysql-connector
neo4j-driver